This patch is a follow-up to r2244 which fixed bugzilla #287260 on some platforms but not on all that we test.
The issue was that cprop_BB did not see that in Add32(t2,t3) the
driving expressions for t2 and t3 were the same. Therefore, the 
Add was not replaced with a shift (which is necessary for proper 
memcheck operation).
So in this patch:

(1) In cprop_BB, when setting up the "env", record *any* assignment
    to a temporary (and not just those that are subject to copy
    propagation).
(2) Pass this env down to fold_Expr and then sameIRExprs. 
(3) Replace sameIRTemps with sameIRExprs and enhance it. Upon 
    encountering an RdTmp, check "env" and recurse into the 
    expression assigned to the temporary.
    As a side, the functions sameIcoU32s and sameIRTempsOrIcoU32s
    and replaced with sameIRExprs.
(4) Add some machinery to monitor frequency and effectiveness of
    sameIRExprs (can be enabled by setting STATS_IROPT).

Hopefully, fixes #287260 for good.


git-svn-id: svn://svn.valgrind.org/vex/trunk@2246 8f6e269a-dfd6-0310-a8e1-e2731360e62c