Bug 369459 - valgrind on arm64 violates the ARMv8 spec (ldxr/stxr)
This implements a fallback LL/SC implementation as described in bug 344524.

The fallback implementation is not enabled by default, and there is no
auto-detection for when it should be used.  To use it, run with the
flag --sim-hints=fallback-llsc.  This commit also allows the existing
MIPS fallback implementation to be enabled with that flag.

VEX side changes:

* priv/main_main.c, pub/libvex.h

  Adds new field guest__use_fallback_LLSC to VexAbiInfo

* pub/libvex_guest_arm64.h priv/guest_arm64_toIR.c

  add front end support, new guest state fields
  guest_LLSC_{SIZE,ADDR,DATA}, also documentation of the scheme

* priv/guest_mips_toIR.c

  allow manual selection of fallback implementation via
  --sim-hints=fallback-llsc

* priv/host_arm64_defs.c priv/host_arm64_defs.h priv/host_arm64_isel.c

  Add support for generating CAS on arm64, as needed by the front end changes



git-svn-id: svn://svn.valgrind.org/vex/trunk@3352 8f6e269a-dfd6-0310-a8e1-e2731360e62c